refactor: add login custom hook
